+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * change inheritance of `User` from `DomainResource` to `AccountDomainResource` (OBJECTS-322)
 * added a reusable compound validation constraint annotation for realms (OBJECTS-568)
+* added default values for `TimelineEntry` (OBJECTS-321)
 * OBJECTS-582 Java Client `UpsertCommand` and `GetCollectionCommand` does not catch `ResourceException`
 * OBJECTS-584 Add method to create ResponseEntity from Result
 * OBJECTS-585 MetadataClient throws ServiceException when looking up by key